requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pyJWT
bcrypt
python-jose
python-multipart